+++ spec/player_rspec.rb
@@ -1,0 +1,1 @@
+require_relative '../lib/players.rb'
